[#163] Relocate HasDetails model to mixins file
The HasDetails abstract model mixin, used for models that require a
details JSONField, has been moved from the base.py file to the mixins.py
file. This reorganizes the model mixins into a dedicated file,
separating them from the base model classes, promoting better code
organization and maintainability.
