+++ apps/api/afbcore/models/base.py
@@ -41,17 +41,3 @@ class BaseAbstractModel(UUIDModel, SoftDeletableModel, TimeStampedModel):
         abstract = True
 
     objects = BaseAbstractModelManager()
-
-
-class HasDetails(models.Model):
-    """
-    A mixin for models that have a details field.
-
-    Fields:
-        - details: JSONField, blank=True, default=dict
-    """
-
-    details = models.JSONField(blank=True, default=dict)
-
-    class Meta:
-        abstract = True
